+++ src/main/java/com/social/backend/repositories/ProjectRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProjectRepo extends JpaRepository<Project, Integer> {
-    List<Project> findByUserId(int userId);
+public interface ProjectRepo extends JpaRepository<Project, Long> {
+    List<Project> findByUserId(Long userId);
 
 }
