+++ src/main/java/com/social/backend/services/ProjectService.java
@@ -3,7 +3,7 @@ package com.social.backend.services;
 import com.social.backend.payloads.ProjectDto;
 
 public interface ProjectService {
-    ProjectDto addProject(Integer userId, ProjectDto projectDto);
-    ProjectDto updateProject(Integer projectId, ProjectDto projectDto);
-    void deleteProject(Integer projectId);
+    ProjectDto addProject(Long userId, ProjectDto projectDto);
+    ProjectDto updateProject(Long projectId, ProjectDto projectDto);
+    void deleteProject(Long projectId);
 }
